reservation completion

+++ src/main/java/com/example/parkinglot/entity/Price.java
@@ -12,7 +12,7 @@ import java.math.BigInteger;
 public class Price {
 
     @Id
-    private int duration;
+    private long duration;
 
     private BigDecimal price;
 }
